pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "pydantic-docs"
description = "Tooling for Pydantic documentation"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "click>=8.0.0",
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.6.0",
    "ruamel.yaml>=0.18.0",
]
# As we provide a CLI, we want to avoid using `importlib.metadata.version()`
# to populate the `__version__` attribute (known to be slow):
dynamic = ["version"]

[project.urls]
Homepage = "https://github.com/pydantic/pydantic-docs"
Source = "https://github.com/pydantic/pydantic-docs"
Changelog = "https://github.com/pydantic/pydantic-docs/blob/main/HISTORY.md"

[project.scripts]
pydantic-docs = "pydantic_docs.cli:cli"

# [project.entry-points."mkdocs.plugins"]
# pydantic-docs = "pydantic_docs.plugin:PydanticPlugin"

[dependency-groups]
dev = [
    "ruff",
    "pyright",
]

[tool.hatch.version]
path = "src/pydantic_docs/__init__.py"

[tool.uv]
default-groups = ["dev"]
required-version = ">=0.8.4"

[tool.ruff]
line-length = 120
target-version = "py39"

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle (E)
    "W",      # pycodestyle (W)
    "F",      # Pyflakes
    "UP",     # pyupgrade
    "I",      # isort
    "PL",     # Pylint
    "RUF",    # Ruff
    "YTT",    # flake8-2020
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "T10",    # flake8-debugger
    "FA",     # flake8-future-annotations
    "PIE",    # flake8-pie
    "PYI",    # flake8-pyi
    "T20",    # flake8-print
    "Q",      # flake8-quotes
    "RSE",    # flake8-raise
    "PTH",    # flake8-use-pathlib
]
ignore = [
    "PLR0912",
]

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"

[tool.ruff.lint.isort]
known-first-party = ["pydantic_docs"]

[tool.pyright]
typeCheckingMode = "standard"
pythonVersion = "3.13"
include = ["src"]
